requirements.txt
discord.py
asyncpg
terminaltables
pytz